ADJ: Move preprocessing funcs from request to preprocess
