+++ request_price.py
@@ -3,7 +3,7 @@ import datetime as dt
 import numpy as np
 import pandas as pd
 
-from datetime import datetime, timedelta
+from preprocess import *
 from ticker import *
 
 DATA = ['Ticker', 'StockSection', 'Date', 'Close']
@@ -16,8 +16,6 @@ class RequestPrice(Ticker):
                  mkt: str = "KOSPI",
                  login_info: str = "login.json"):
         super().__init__(mkt, login_info)
-        if not self.config_connect:
-            self.process_connect(login_info=login_info)
 
         self.start = start
         if end == 'TODAY':
@@ -25,10 +23,11 @@ class RequestPrice(Ticker):
                         dt.timedelta(days=1)).strftime("%Y%m%d")
         else:
             self.end = end
+        self.preprocess = PreProcess(self.start, self.end)
 
         self.obj = self.objStockChart
 
-        self.ticker = self.get_ticker_info
+        self.ticker = self.get_ticker_info[:5]
         self.ROW = list(range(len(DATA)))
         self.ROWS = []
         self.res = None
@@ -96,38 +95,6 @@ class RequestPrice(Ticker):
                 self.ROWS.append(list(self.ROW))
         return self.ROWS
 
-    @staticmethod
-    def pp_extract_dates(df: pd.DataFrame) -> list:
-        """
-        Extract dates for time-series data columns.
-        """
-        dates = df['Date'].tolist()
-        dates_sliced = []
-        current = [dates[0]]
-
-        for i in range(1, len(dates)):
-            if dates[i] > dates[i-1]:
-                dates_sliced.append(current)
-                current = [dates[i]]
-            else:
-                current.append(dates[i])
-
-        dates_sliced.append(current)
-        res = max(dates_sliced, key=len)
-        return res
-
-    @staticmethod
-    def pp_append_nans(df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Append NaN values for companies listed after start date.
-        (i.e. Listed between start date and end date.)
-        """
-        res = df.groupby('Ticker')['Close'].agg(list).reset_index()
-        max_date = res['Close'].apply(len).max()
-        res['Close'] = res['Close'].apply(
-            lambda x: x + [np.nan] * (max_date - len(x)))
-        return res
-
     @classmethod
     def pp_dt_storage(cls, start, end):
         """
@@ -140,8 +107,8 @@ class RequestPrice(Ticker):
         Get time-series price data.
         """
         df = pd.DataFrame(self.get_request, columns=DATA)
-        dates = self.pp_extract_dates(df)
-        temp = self.pp_append_nans(df)
+        dates = self.preprocess.pp_extract_dates(df)
+        temp = self.preprocess.pp_append_nans(df)
         self.res = pd.DataFrame(temp['Close'].tolist(),
                                 index=temp['Ticker'],
                                 columns=dates).T[::-1]
@@ -155,29 +122,12 @@ class RequestSplit(RequestPrice):
                  mkt: str = "KOSPI",
                  login_info: str = "login.json"):
         super().__init__(start, end, mkt, login_info)
-        self.time_interval = timedelta(days=252 * 3)
-
-    def pp_time_diff(self) -> list:
-        """
-        Limit time difference to 3-year business days.
-        Slice time interval.
-        """
-        start_dt = datetime.strptime(self.start, "%Y%m%d")
-        end_dt = datetime.strptime(self.end, "%Y%m%d")
-
-        res_dt = []
-        while start_dt < end_dt:
-            next_dt = start_dt + self.time_interval
-            res_dt.append([datetime.strftime(start_dt, "%Y%m%d"),
-                           datetime.strftime(min(next_dt, end_dt), "%Y%m%d")])
-            start_dt = next_dt + timedelta(days=1)
-        return res_dt
 
     def get_split_price(self) -> pd.DataFrame:
         """
         Get time-series price data.
         """
-        res_dt = self.pp_time_diff()
+        res_dt = self.preprocess.pp_time_diff()
 
         temp = []
         for start, end in res_dt:
@@ -190,4 +140,5 @@ class RequestSplit(RequestPrice):
 
 
 # TODO: start date에 존재하지 않는 기업 배제: 추가 방안?
+# TODO: StockSectionKind 기반 Ticker 전송
 # TODO: task kill issue
